requirements.txt
Flask
requests
openai
python-pptx
SQLAlchemy
